separated glm and rsa scripts
